feat: add tabs for shell code snippets
