+++ dotnet/docs/browsers.mdx
@@ -120,34 +120,76 @@ du -hs ~/Library/Caches/ms-playwright/*
 
 You can override default behavior using environment variables. When installing Playwright, ask it to download browsers into a specific location:
 
+<Tabs
+  groupId="bash-flavor"
+  defaultValue="bash"
+  values={[
+    {label: 'Bash', value: 'bash'},
+    {label: 'PowerShell', value: 'powershell'},
+    {label: 'Batch', value: 'batch'}
+  ]
+}>
+<TabItem value="bash">
+
 ```bash
-# Linux/macOS
 PLAYWRIGHT_BROWSERS_PATH=$HOME/pw-browsers 
 playwright install
+```
 
-# Windows with cmd.exe
-set PLAYWRIGHT_BROWSERS_PATH=%USERPROFILE%\pw-browsers
-playwright install
+</TabItem>
+<TabItem value="powershell">
 
-# Windows with PowerShell
+```powershell
 $env:PLAYWRIGHT_BROWSERS_PATH="$env:USERPROFILE\pw-browsers"
 playwright install
 ```
+
+</TabItem>
+<TabItem value="batch">
+
+```batch
+set PLAYWRIGHT_BROWSERS_PATH=%USERPROFILE%\pw-browsers
+playwright install
+```
+
+</TabItem>
+</Tabs>
 
 When running Playwright scripts, ask it to search for browsers in a shared location.
 
+<Tabs
+  groupId="bash-flavor"
+  defaultValue="bash"
+  values={[
+    {label: 'Bash', value: 'bash'},
+    {label: 'PowerShell', value: 'powershell'},
+    {label: 'Batch', value: 'batch'}
+  ]
+}>
+<TabItem value="bash">
+
 ```bash
-# Linux/macOS
 PLAYWRIGHT_BROWSERS_PATH=$HOME/pw-browsers dotnet test
+```
 
-# Windows with cmd.exe
-set PLAYWRIGHT_BROWSERS_PATH=%USERPROFILE%\pw-browsers
-dotnet test
+</TabItem>
+<TabItem value="powershell">
 
-# Windows with PowerShell
+```powershell
 $env:PLAYWRIGHT_BROWSERS_PATH="$env:USERPROFILE\pw-browsers"
 dotnet test
 ```
+
+</TabItem>
+<TabItem value="batch">
+
+```batch
+set PLAYWRIGHT_BROWSERS_PATH=%USERPROFILE%\pw-browsers
+dotnet test
+```
+
+</TabItem>
+</Tabs>
 
 Playwright keeps track of packages that need those browsers and will garbage collect them as you update Playwright to the newer versions.
 
@@ -160,21 +202,42 @@ Developers can opt-in in this mode via exporting `PLAYWRIGHT_BROWSERS_PATH=$HOME
 
 You can opt into the hermetic install and place binaries in the local folder:
 
+<Tabs
+  groupId="bash-flavor"
+  defaultValue="bash"
+  values={[
+    {label: 'Bash', value: 'bash'},
+    {label: 'PowerShell', value: 'powershell'},
+    {label: 'Batch', value: 'batch'}
+  ]
+}>
+<TabItem value="bash">
+
 ```bash
-# Linux/macOS
 # Places binaries to node_modules/@playwright/test
 PLAYWRIGHT_BROWSERS_PATH=0 npx playwright install
+```
 
-# Windows with cmd.exe
-# Places binaries to node_modules\@playwright\test
-set PLAYWRIGHT_BROWSERS_PATH=0
-npx playwright install
+</TabItem>
+<TabItem value="powershell">
 
-# Windows with PowerShell
+```powershell
 # Places binaries to node_modules\@playwright\test
 $env:PLAYWRIGHT_BROWSERS_PATH=0
 npx playwright install
 ```
+
+</TabItem>
+<TabItem value="batch">
+
+```batch
+# Places binaries to node_modules\@playwright\test
+set PLAYWRIGHT_BROWSERS_PATH=0
+npx playwright install
+```
+
+</TabItem>
+</Tabs>
 
 ## Install behind a firewall or a proxy
 
@@ -182,18 +245,39 @@ By default, Playwright downloads browsers from Microsoft CDN.
 
 Sometimes companies maintain an internal proxy that blocks direct access to the public resources. In this case, Playwright can be configured to download browsers via a proxy server.
 
+<Tabs
+  groupId="bash-flavor"
+  defaultValue="bash"
+  values={[
+    {label: 'Bash', value: 'bash'},
+    {label: 'PowerShell', value: 'powershell'},
+    {label: 'Batch', value: 'batch'}
+  ]
+}>
+<TabItem value="bash">
+
 ```bash
-# Linux/macOS
 HTTPS_PROXY=https://192.0.2.1 playwright install
+```
 
-# Windows with cmd.exe
-set HTTPS_PROXY=https://192.0.2.1
-playwright install
+</TabItem>
+<TabItem value="powershell">
 
-# Windows with PowerShell
+```powershell
 $env:HTTPS_PROXY="https://192.0.2.1"
 playwright install
 ```
+
+</TabItem>
+<TabItem value="batch">
+
+```batch
+set HTTPS_PROXY=https://192.0.2.1
+playwright install
+```
+
+</TabItem>
+</Tabs>
 
 ## Download from artifact repository
 
@@ -201,18 +285,39 @@ By default, Playwright downloads browsers from Microsoft CDN.
 
 Sometimes companies maintain an internal artifact repository to host browser binaries. In this case, Playwright can be configured to download from a custom location using the `PLAYWRIGHT_DOWNLOAD_HOST` env variable.
 
+<Tabs
+  groupId="bash-flavor"
+  defaultValue="bash"
+  values={[
+    {label: 'Bash', value: 'bash'},
+    {label: 'PowerShell', value: 'powershell'},
+    {label: 'Batch', value: 'batch'}
+  ]
+}>
+<TabItem value="bash">
+
 ```bash
-# Linux/macOS
 PLAYWRIGHT_DOWNLOAD_HOST=192.0.2.1 playwright install
+```
 
-# Windows with cmd.exe
-set PLAYWRIGHT_DOWNLOAD_HOST=192.0.2.1
-playwright install
+</TabItem>
+<TabItem value="powershell">
 
-# Windows with PowerShell
+```powershell
 $env:PLAYWRIGHT_DOWNLOAD_HOST="192.0.2.1"
 playwright install
 ```
+
+</TabItem>
+<TabItem value="batch">
+
+```batch
+set PLAYWRIGHT_DOWNLOAD_HOST=192.0.2.1
+playwright install
+```
+
+</TabItem>
+</Tabs>
 
 It is also possible to use a per-browser download hosts using `PLAYWRIGHT_CHROMIUM_DOWNLOAD_HOST`, `PLAYWRIGHT_FIREFOX_DOWNLOAD_HOST` and `PLAYWRIGHT_WEBKIT_DOWNLOAD_HOST` env variables that take precedence over `PLAYWRIGHT_DOWNLOAD_HOST`.
 
@@ -227,18 +332,39 @@ In certain cases, it is desired to avoid browser downloads altogether because br
 
 This can be done by setting `PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD` variable before installation.
 
+<Tabs
+  groupId="bash-flavor"
+  defaultValue="bash"
+  values={[
+    {label: 'Bash', value: 'bash'},
+    {label: 'PowerShell', value: 'powershell'},
+    {label: 'Batch', value: 'batch'}
+  ]
+}>
+<TabItem value="bash">
+
 ```bash
-# Linux/macOS
 PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD=1 playwright install
+```
 
-# Windows with cmd.exe
-set PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD=1
-playwright install
+</TabItem>
+<TabItem value="powershell">
 
-# Windows with PowerShell
+```powershell
 $env:PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD=1
 playwright install
 ```
+
+</TabItem>
+<TabItem value="batch">
+
+```batch
+set PLAYWRIGHT_SKIP_BROWSER_DOWNLOAD=1
+playwright install
+```
+
+</TabItem>
+</Tabs>
 
 ## Stale browser removal
