+++ dotnet/docs/debug.mdx
@@ -107,19 +107,39 @@ Setup [`launch.json` configuration](https://code.visualstudio.com/docs/nodejs/no
 
 Playwright supports verbose logging with the `DEBUG` environment variable.
 
+<Tabs
+  groupId="bash-flavor"
+  defaultValue="bash"
+  values={[
+    {label: 'Bash', value: 'bash'},
+    {label: 'PowerShell', value: 'powershell'},
+    {label: 'Batch', value: 'batch'}
+  ]
+}>
+<TabItem value="bash">
+
 ```bash
-# Linux/macOS
 DEBUG=pw:api dotnet run
+```
 
-# Windows with cmd.exe
-set DEBUG=pw:api
-dotnet run
+</TabItem>
+<TabItem value="powershell">
 
-# Windows with PowerShell
+```powershell
 $env:DEBUG="pw:api"
 dotnet run
 ```
 
+</TabItem>
+<TabItem value="batch">
+
+```batch
+set DEBUG=pw:api
+dotnet run
+```
+
+</TabItem>
+</Tabs>
 
 [Accessibility]: ./api/class-accessibility.mdx "Accessibility"
 [Browser]: ./api/class-browser.mdx "Browser"
