+++ nodejs/docs/api/class-fixtures.mdx
@@ -19,7 +19,7 @@ Playwright Test looks at each test declaration, analyses the set of fixtures the
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test('basic test', async ({ page }) => {
@@ -73,7 +73,7 @@ Name of the browser that runs tests. Defaults to `'chromium'`. Useful to [annota
 }>
 <TabItem value="ts">
 
-```js
+```ts
 test('skip this test in Firefox', async ({ page, browserName }) => {
   test.skip(browserName === 'firefox', 'Still working on it');
   // ...
@@ -119,7 +119,7 @@ This is the most common fixture used in a test.
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test('basic test', async ({ page }) => {
