+++ nodejs/docs/api/class-reporter.mdx
@@ -19,7 +19,7 @@ You can create a custom reporter by implementing a class with some of the report
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // my-awesome-reporter.ts
 import { Reporter } from '@playwright/test/reporter';
 
@@ -87,7 +87,7 @@ Now use this reporter with [testConfig.reporter](./api/class-testconfig.mdx#test
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
