+++ nodejs/docs/api/class-test.mdx
@@ -17,7 +17,7 @@ Playwright Test provides a `test` function to declare tests and [`expect` functi
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test('basic test', async ({ page }) => {
@@ -85,7 +85,7 @@ Declares a test.
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test('basic test', async ({ page }) => {
@@ -139,7 +139,7 @@ Declares a `beforeAll` hook that is executed once before all tests. When called 
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // example.spec.ts
 import { test, expect } from '@playwright/test';
 
@@ -197,7 +197,7 @@ Declares a `beforeEach` hook that is executed before each test. When called in t
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // example.spec.ts
 import { test, expect } from '@playwright/test';
 
@@ -250,7 +250,7 @@ Declares a group of tests.
 }>
 <TabItem value="ts">
 
-```js
+```ts
 test.describe('two tests', () => {
   test('one', async ({ page }) => {
     // ...
@@ -297,7 +297,7 @@ Declares a focused group of tests. If there are some focused tests or suites, al
 }>
 <TabItem value="ts">
 
-```js
+```ts
 test.describe.only('focused group', () => {
   test('in the focused group', async ({ page }) => {
     // This test will run
@@ -342,7 +342,7 @@ Declares a group of tests that could be run in parallel. By default, tests in a 
 }>
 <TabItem value="ts">
 
-```js
+```ts
 test.describe.parallel('group', () => {
   test('runs in parallel 1', async ({ page }) => {
   });
@@ -396,7 +396,7 @@ Using serial is not recommended. It is usually better to make your tests isolate
 }>
 <TabItem value="ts">
 
-```js
+```ts
 test.describe.serial('group', () => {
   test('runs first', async ({ page }) => {
   });
@@ -441,7 +441,7 @@ Using serial is not recommended. It is usually better to make your tests isolate
 }>
 <TabItem value="ts">
 
-```js
+```ts
 test.describe.serial.only('group', () => {
   test('runs first', async ({ page }) => {
   });
@@ -484,7 +484,7 @@ Unconditional fail:
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test('not yet ready', async ({ page }) => {
@@ -520,7 +520,7 @@ Conditional fail a test with an optional description:
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test('fail in WebKit', async ({ page, browserName }) => {
@@ -556,7 +556,7 @@ Conditional fail for all tests in a file or [test.describe(title, callback)](./a
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test.fail(({ browserName }) => browserName === 'webkit');
@@ -607,7 +607,7 @@ Unconditional fixme:
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test('not yet ready', async ({ page }) => {
@@ -643,7 +643,7 @@ Conditional fixme a test with an optional description:
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test('fixme in WebKit', async ({ page, browserName }) => {
@@ -679,7 +679,7 @@ Conditional fixme for all tests in a file or [test.describe(title, callback)](./
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test.fixme(({ browserName }) => browserName === 'webkit');
@@ -723,7 +723,7 @@ test('fixme in WebKit 2', async ({ page }) => {
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test.beforeEach(async ({ page }) => {
@@ -764,7 +764,7 @@ Declares a focused test. If there are some focused tests or suites, all of them 
 }>
 <TabItem value="ts">
 
-```js
+```ts
 test.only('focus this test', async ({ page }) => {
   // Run only focused tests in the entire project.
 });
@@ -798,7 +798,7 @@ Changes the timeout for the test.
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test('very slow test', async ({ page }) => {
@@ -834,7 +834,7 @@ Changing timeout from a slow hook:
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test.beforeEach(async ({ page }, testInfo) => {
@@ -877,7 +877,7 @@ Declares a skipped test, similarly to [test(title, testFunction)](./api/class-te
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test.skip('broken test', async ({ page }) => {
@@ -914,7 +914,7 @@ Unconditionally skip a test. Test is immediately aborted when you call [test.ski
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test('skipped test', async ({ page }) => {
@@ -950,7 +950,7 @@ Unconditionally skip all tests in a file or [test.describe(title, callback)](./a
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test.skip();
@@ -999,7 +999,7 @@ Conditionally skip a test with an optional description.
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test('skip in WebKit', async ({ page, browserName }) => {
@@ -1035,7 +1035,7 @@ Skip from [test.beforeEach(hookFunction)](./api/class-test.mdx#test-before-each)
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test.beforeEach(async ({ page }) => {
@@ -1076,7 +1076,7 @@ Conditionally skips all tests in a file or [test.describe(title, callback)](./ap
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test.skip(({ browserName }) => browserName === 'webkit');
@@ -1127,7 +1127,7 @@ Unconditional slow:
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test('slow test', async ({ page }) => {
@@ -1163,7 +1163,7 @@ Conditional slow a test with an optional description:
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test('slow in WebKit', async ({ page, browserName }) => {
@@ -1199,7 +1199,7 @@ Conditional slow for all tests in a file or [test.describe(title, callback)](./a
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test.slow(({ browserName }) => browserName === 'webkit');
@@ -1248,7 +1248,7 @@ Declares a test step.
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test('test', async ({ page }) => {
@@ -1290,7 +1290,7 @@ Specifies options or fixtures to use in a single test file or a [test.describe(t
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test.use({ locale: 'en-US' });
@@ -1328,7 +1328,7 @@ It is also possible to override a fixture by providing a function.
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test.use({
