+++ nodejs/docs/api/class-testconfig.mdx
@@ -19,7 +19,7 @@ Playwright Test supports running multiple test projects at the same time. Projec
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
 
@@ -109,7 +109,7 @@ Learn more about [global setup and teardown](./test-advanced.mdx#global-setup-an
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig, devices } from '@playwright/test';
 
@@ -194,7 +194,7 @@ Here is an example that uses [testInfo.outputPath(pathSegments)](./api/class-tes
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 import fs from 'fs';
 
@@ -276,7 +276,7 @@ Learn more in the [reporters guide](./test-reporters.mdx).
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
 
@@ -369,7 +369,7 @@ Global options for all tests, for example [testOptions.browserName](./api/class-
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
