+++ nodejs/docs/api/class-testinfo.mdx
@@ -17,7 +17,7 @@ import TabItem from '@theme/TabItem';
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test('basic test', async ({ page }, testInfo) => {
@@ -101,7 +101,7 @@ Returns a path inside the [testInfo.outputDir](./api/class-testinfo.mdx#test-inf
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 import fs from 'fs';
 
@@ -148,7 +148,7 @@ Timeout is usually specified in the [configuration file](./test-configuration.md
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test.beforeEach(async ({ page }, testInfo) => {
@@ -223,7 +223,7 @@ The list of files or buffers attached to the current test. Some reporters show t
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test('basic test', async ({ page }, testInfo) => {
@@ -297,7 +297,7 @@ Expected status is usually compared with the actual [testInfo.status](./api/clas
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test.afterEach(async ({}, testInfo) => {
@@ -378,7 +378,7 @@ Status is usually compared with the [testInfo.expectedStatus](./api/class-testin
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test.afterEach(async ({}, testInfo) => {
@@ -427,7 +427,7 @@ Timeout in milliseconds for the currently running test. Zero means no timeout. T
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test.beforeEach(async ({ page }, testInfo) => {
