+++ nodejs/docs/api/class-testoptions.mdx
@@ -19,7 +19,7 @@ These options are usually provided in the [configuration file](./test-configurat
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { PlaywrightTestConfig } from '@playwright/test';
 const config: PlaywrightTestConfig = {
   use: {
@@ -66,7 +66,7 @@ Alternatively, with [test.use(options)](./api/class-test.mdx#test-use) you can o
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // example.spec.ts
 import { test, expect } from '@playwright/test';
 
@@ -162,7 +162,7 @@ Name of the browser that runs tests. Defaults to `'chromium'`. Most of the time 
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig, devices } from '@playwright/test';
