+++ nodejs/docs/api/class-testproject.mdx
@@ -21,7 +21,7 @@ Here is an example configuration that runs every test in Chromium, Firefox and W
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig, devices } from '@playwright/test';
 
@@ -174,7 +174,7 @@ Here is an example that uses [testInfo.outputPath(pathSegments)](./api/class-tes
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 import fs from 'fs';
 
@@ -227,7 +227,7 @@ Each project can use a different directory. Here is an example that runs smoke t
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
 
@@ -351,7 +351,7 @@ Options for all tests in this project, for example [testOptions.browserName](./a
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
