+++ nodejs/docs/api/class-workerinfo.mdx
@@ -17,7 +17,7 @@ import TabItem from '@theme/TabItem';
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test.beforeAll(async ({ browserName }, workerInfo) => {
