+++ nodejs/docs/browsers.mdx
@@ -49,7 +49,7 @@ While Playwright can download and use the recent Chromium build, it can operate 
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { PlaywrightTestConfig } from '@playwright/test';
 const config: PlaywrightTestConfig = {
   use: {
@@ -78,7 +78,7 @@ module.exports = config;
 </TabItem>
 <TabItem value="library">
 
-```js
+```library
 const { chromium } = require('playwright');
 const browser = await chromium.launch({
   channel: 'chrome' // or 'msedge', 'chrome-beta', 'msedge-beta', 'msedge-dev', etc.
@@ -126,33 +126,75 @@ du -hs ~/Library/Caches/ms-playwright/*
 
 You can override default behavior using environment variables. When installing Playwright, ask it to download browsers into a specific location:
 
+<Tabs
+  groupId="bash-flavor"
+  defaultValue="bash"
+  values={[
+    {label: 'Bash', value: 'bash'},
+    {label: 'PowerShell', value: 'powershell'},
+    {label: 'Batch', value: 'batch'}
+  ]
+}>
+<TabItem value="bash">
+
 ```bash
-# Linux/macOS
 PLAYWRIGHT_BROWSERS_PATH=$HOME/pw-browsers npx playwright install
+```
 
-# Windows with cmd.exe
-set PLAYWRIGHT_BROWSERS_PATH=%USERPROFILE%\pw-browsers
-npx playwright install
+</TabItem>
+<TabItem value="powershell">
 
-# Windows with PowerShell
+```powershell
 $env:PLAYWRIGHT_BROWSERS_PATH="$env:USERPROFILE\pw-browsers"
 npx playwright install
 ```
+
+</TabItem>
+<TabItem value="batch">
+
+```batch
+set PLAYWRIGHT_BROWSERS_PATH=%USERPROFILE%\pw-browsers
+npx playwright install
+```
+
+</TabItem>
+</Tabs>
 
 When running Playwright scripts, ask it to search for browsers in a shared location.
 
+<Tabs
+  groupId="bash-flavor"
+  defaultValue="bash"
+  values={[
+    {label: 'Bash', value: 'bash'},
+    {label: 'PowerShell', value: 'powershell'},
+    {label: 'Batch', value: 'batch'}
+  ]
+}>
+<TabItem value="bash">
+
 ```bash
-# Linux/macOS
 PLAYWRIGHT_BROWSERS_PATH=$HOME/pw-browsers npx playwright test
+```
 
-# Windows with cmd.exe
-set PLAYWRIGHT_BROWSERS_PATH=%USERPROFILE%\pw-browsers
-npx playwright test
+</TabItem>
+<TabItem value="powershell">
 
-# Windows with PowerShell
+```powershell
 $env:PLAYWRIGHT_BROWSERS_PATH="$env:USERPROFILE\pw-browsers"
 npx playwright test
 ```
+
+</TabItem>
+<TabItem value="batch">
+
+```batch
+set PLAYWRIGHT_BROWSERS_PATH=%USERPROFILE%\pw-browsers
+npx playwright test
+```
+
+</TabItem>
+</Tabs>
 
 Playwright keeps track of packages that need those browsers and will garbage collect them as you update Playwright to the newer versions.
 
@@ -165,21 +207,42 @@ Developers can opt-in in this mode via exporting `PLAYWRIGHT_BROWSERS_PATH=$HOME
 
 You can opt into the hermetic install and place binaries in the local folder:
 
+<Tabs
+  groupId="bash-flavor"
+  defaultValue="bash"
+  values={[
+    {label: 'Bash', value: 'bash'},
+    {label: 'PowerShell', value: 'powershell'},
+    {label: 'Batch', value: 'batch'}
+  ]
+}>
+<TabItem value="bash">
+
 ```bash
-# Linux/macOS
 # Places binaries to node_modules/@playwright/test
 PLAYWRIGHT_BROWSERS_PATH=0 npx playwright install
+```
 
-# Windows with cmd.exe
-# Places binaries to node_modules\@playwright\test
-set PLAYWRIGHT_BROWSERS_PATH=0
-npx playwright install
+</TabItem>
+<TabItem value="powershell">
 
-# Windows with PowerShell
+```powershell
 # Places binaries to node_modules\@playwright\test
 $env:PLAYWRIGHT_BROWSERS_PATH=0
 npx playwright install
 ```
+
+</TabItem>
+<TabItem value="batch">
+
+```batch
+# Places binaries to node_modules\@playwright\test
+set PLAYWRIGHT_BROWSERS_PATH=0
+npx playwright install
+```
+
+</TabItem>
+</Tabs>
 
 ## Install behind a firewall or a proxy
