+++ nodejs/docs/debug.mdx
@@ -66,18 +66,39 @@ Using `PWDEBUG=console` will configure the browser for debugging in Developer to
 * **Disables timeout**: Sets default timeout to 0 (= no timeout)
 * **Console helper**: Configures a `playwright` object in the browser to generate and highlight [Playwright selectors](./selectors.mdx). This can be used to verify text or composite selectors.
 
+<Tabs
+  groupId="bash-flavor"
+  defaultValue="bash"
+  values={[
+    {label: 'Bash', value: 'bash'},
+    {label: 'PowerShell', value: 'powershell'},
+    {label: 'Batch', value: 'batch'}
+  ]
+}>
+<TabItem value="bash">
+
 ```bash
-# Linux/macOS
 PWDEBUG=console npm run test
+```
 
-# Windows with cmd.exe
-set PWDEBUG=console
-npm run test
+</TabItem>
+<TabItem value="powershell">
 
-# Windows with PowerShell
-$env:PWDEBUG="console"
+```powershell
+env:PWDEBUG="console"
 npm run test
 ```
+
+</TabItem>
+<TabItem value="batch">
+
+```batch
+set PWDEBUG=console
+npm run test
+```
+
+</TabItem>
+</Tabs>
 
 ## Selectors in Developer Tools Console
 
@@ -112,19 +133,39 @@ Setup [`launch.json` configuration](https://code.visualstudio.com/docs/nodejs/no
 
 Playwright supports verbose logging with the `DEBUG` environment variable.
 
+<Tabs
+  groupId="bash-flavor"
+  defaultValue="bash"
+  values={[
+    {label: 'Bash', value: 'bash'},
+    {label: 'PowerShell', value: 'powershell'},
+    {label: 'Batch', value: 'batch'}
+  ]
+}>
+<TabItem value="bash">
+
 ```bash
-# Linux/macOS
 DEBUG=pw:api npm run test
+```
 
-# Windows with cmd.exe
-set DEBUG=pw:api
-npm run test
+</TabItem>
+<TabItem value="powershell">
 
-# Windows with PowerShell
+```powershell
 $env:DEBUG="pw:api"
 npm run test
 ```
 
+</TabItem>
+<TabItem value="batch">
+
+```batch
+set DEBUG=pw:api
+npm run test
+```
+
+</TabItem>
+</Tabs>
 
 [Accessibility]: ./api/class-accessibility.mdx "Accessibility"
 [Android]: ./api/class-android.mdx "Android"
