+++ nodejs/docs/intro.mdx
@@ -53,7 +53,7 @@ Create `tests/foo.spec.js` (or `tests/foo.spec.ts` for TypeScript) to define you
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test('basic test', async ({ page }) => {
@@ -123,7 +123,7 @@ Here is a quick example of using them:
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // example.spec.ts
 import { test, expect } from '@playwright/test';
 
@@ -214,7 +214,7 @@ You noticed an argument `{ page }` that the test above has access to:
 }>
 <TabItem value="ts">
 
-```js
+```ts
 test('basic test', async ({ page }) => {
   ...
 ```
@@ -255,7 +255,7 @@ You can use `test.beforeAll` and `test.afterAll` hooks to set up and tear down r
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // example.spec.ts
 import { test, expect } from '@playwright/test';
 
@@ -386,7 +386,7 @@ Create `playwright.config.ts` (or `playwright.config.js`) to configure your test
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig, devices } from '@playwright/test';
