+++ nodejs/docs/test-advanced.mdx
@@ -76,7 +76,7 @@ Consider an example where we run a new http server per worker process, and use `
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // my-test.ts
 import { test as base } from '@playwright/test';
 import * as http from 'http';
@@ -167,7 +167,7 @@ Here is an example test that saves some information:
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // example.spec.ts
 import { test } from '@playwright/test';
 
@@ -209,7 +209,7 @@ Here is an example fixture that automatically saves debug logs when the test fai
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // my-test.ts
 import * as debug from 'debug';
 import * as fs from 'fs';
@@ -277,7 +277,7 @@ The port gets then passed over to Playwright as a `baseURL` when creating the co
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
 const config: PlaywrightTestConfig = {
@@ -324,7 +324,7 @@ Now you can use a relative path when navigating the page, or use `baseURL` fixtu
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // test.spec.ts
 import { test } from '@playwright/test';
 test('test', async ({ page, baseURL }) => {
@@ -376,7 +376,7 @@ Here is a global setup example that authenticates once and reuses authentication
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // global-setup.ts
 import { chromium, FullConfig } from '@playwright/test';
 
@@ -430,7 +430,7 @@ Specify `globalSetup`, `baseURL` and `storageState` in the configuration file.
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
 
@@ -476,7 +476,7 @@ Tests start already authenticated because we specify `storageState` that was pop
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test } from '@playwright/test';
 
 test('test', async ({ page }) => {
@@ -516,7 +516,7 @@ To make use of this feature, we will declare an "option fixture" for the backend
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // my-test.ts
 import { test as base } from '@playwright/test';
 import { startBackend } from './my-backend';
@@ -575,7 +575,7 @@ We can use our fixtures in the test.
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // example.spec.ts
 import { test } from './my-test';
 
@@ -627,7 +627,7 @@ Now, we can run test in multiple configurations by using projects.
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
 import { TestOptions } from './my-test';
@@ -704,7 +704,7 @@ In this example we add a custom `toBeWithinRange` function in the configuration 
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { expect, PlaywrightTestConfig } from '@playwright/test';
 
@@ -771,7 +771,7 @@ Now we can use `toBeWithinRange` in the test.
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // example.spec.ts
 import { test, expect } from '@playwright/test';
