+++ nodejs/docs/test-annotations.mdx
@@ -38,7 +38,7 @@ You can focus some tests. When there are focused tests, only these tests run.
 }>
 <TabItem value="ts">
 
-```js
+```ts
 test.only('focus this test', async ({ page }) => {
   // Run only focused tests in the entire project.
 });
@@ -70,7 +70,7 @@ Mark a test as skipped.
 }>
 <TabItem value="ts">
 
-```js
+```ts
 test.skip('skip this test', async ({ page }) => {
   // This test is not run
 });
@@ -102,7 +102,7 @@ You can skip certain test based on the condition.
 }>
 <TabItem value="ts">
 
-```js
+```ts
 test('skip this test', async ({ page, browserName }) => {
   test.skip(browserName === 'firefox', 'Still working on it');
 });
@@ -134,7 +134,7 @@ You can group tests to give them a logical name or to scope before/after hooks t
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test.describe('two tests', () => {
@@ -182,7 +182,7 @@ Sometimes you want to tag your tests as `@fast` or `@slow` and only run the test
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test, expect } from '@playwright/test';
 
 test('Test login page @fast', async ({ page }) => {
@@ -238,7 +238,7 @@ For example, you can run a group of tests just in Chromium by passing a callback
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // example.spec.ts
 
 test.describe('chromium only', () => {
@@ -298,7 +298,7 @@ To avoid running `beforeEach` hooks, you can put annotations in the hook itself.
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // example.spec.ts
 
 test.beforeEach(async ({ page }) => {
