+++ nodejs/docs/test-auth.mdx
@@ -28,7 +28,7 @@ This is the simplest way where each test signs in inside the `beforeEach` hook. 
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test } from '@playwright/test';
 
 test.beforeEach(async ({ page }) => {
@@ -94,7 +94,7 @@ Create a new global setup script:
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // global-setup.ts
 import { chromium, FullConfig } from '@playwright/test';
 
@@ -148,7 +148,7 @@ Register global setup script in the Playwright configuration file:
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
 
@@ -194,7 +194,7 @@ Tests start already authenticated because we specify `storageState` that was pop
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test } from '@playwright/test';
 
 test('test', async ({ page }) => {
@@ -234,7 +234,7 @@ Sometimes you have more than one signed-in user in your end to end tests. You ca
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // global-setup.ts
 import { chromium, FullConfig } from '@playwright/test';
 
@@ -288,7 +288,7 @@ After that you can specify the user to use for each test file or each test group
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test } from '@playwright/test';
 
 test.use({ storageState: 'adminStorageState.json' });
@@ -344,7 +344,7 @@ Although discouraged, sometimes it is necessary to sacrifice the isolation and r
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // example.spec.ts
 
 import { test, Page } from '@playwright/test';
