+++ nodejs/docs/test-configuration.mdx
@@ -39,7 +39,7 @@ Create `playwright.config.js` (or `playwright.config.ts`) and specify options in
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { PlaywrightTestConfig } from '@playwright/test';
 const config: PlaywrightTestConfig = {
   use: {
@@ -100,7 +100,7 @@ With [test.use(options)](./api/class-test.mdx#test-use) you can override some op
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // example.spec.ts
 import { test, expect } from '@playwright/test';
 
@@ -142,7 +142,7 @@ The same works inside describe.
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // example.spec.ts
 import { test, expect } from '@playwright/test';
 
@@ -198,7 +198,7 @@ These are commonly used options for various scenarios. You usually set them glob
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { PlaywrightTestConfig } from '@playwright/test';
 const config: PlaywrightTestConfig = {
   use: {
@@ -247,7 +247,7 @@ Here is an example configuration that runs tests in "Pixel 4" and "iPhone 11" em
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig, devices } from '@playwright/test';
 
@@ -334,7 +334,7 @@ You can specify options separately instead of using predefined devices. There ar
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { PlaywrightTestConfig } from '@playwright/test';
 const config: PlaywrightTestConfig = {
   use: {
@@ -391,7 +391,7 @@ You don't have to configure anything to mock network requests. Just define a cus
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // example.spec.ts
 import { test, expect } from '@playwright/test';
 
@@ -439,7 +439,7 @@ Alternatively, you can use [page.route(url, handler[, options])](./api/class-pag
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // example.spec.ts
 import { test, expect } from '@playwright/test';
 
@@ -490,7 +490,7 @@ Screenshots will appear in the test output directory, typically `test-results`.
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { PlaywrightTestConfig } from '@playwright/test';
 const config: PlaywrightTestConfig = {
   use: {
@@ -539,7 +539,7 @@ Video files will appear in the test output directory, typically `test-results`.
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { PlaywrightTestConfig } from '@playwright/test';
 const config: PlaywrightTestConfig = {
   use: {
@@ -588,7 +588,7 @@ Trace files will appear in the test output directory, typically `test-results`.
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { PlaywrightTestConfig } from '@playwright/test';
 const config: PlaywrightTestConfig = {
   use: {
@@ -631,7 +631,7 @@ Any options accepted by [browserType.launch([options])](./api/class-browsertype.
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { PlaywrightTestConfig } from '@playwright/test';
 const config: PlaywrightTestConfig = {
   use: {
@@ -692,7 +692,7 @@ You can specify these options in the configuration file. Note that testing optio
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
 
@@ -768,7 +768,7 @@ To specify different options per browser, for example command line arguments for
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
