+++ nodejs/docs/test-fixtures.mdx
@@ -64,7 +64,7 @@ test.describe('todo tests', () => {
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // example.spec.ts
 import { test as base } from '@playwright/test';
 import { TodoPage } from './todo-page';
@@ -144,7 +144,7 @@ Test fixtures are set up for each test. Consider the following test file:
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // hello.spec.ts
 import test from './hello';
 
@@ -190,7 +190,7 @@ Here is how test fixtures are declared and defined. Fixtures can use other fixtu
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // hello.ts
 import { test as base } from '@playwright/test';
 
@@ -268,7 +268,7 @@ Here is how the test looks:
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // express.spec.ts
 import test from './express-test';
 import fetch from 'node-fetch';
@@ -318,7 +318,7 @@ And here is how fixtures are declared and defined:
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // express-test.ts
 import { test as base } from '@playwright/test';
 import express from 'express';
@@ -437,7 +437,7 @@ In addition to creating your own fixtures, you can also override existing fixtur
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test as base } from '@playwright/test';
 
 export const test = base.extend({
@@ -477,7 +477,7 @@ Notice that in this example, the `page` fixture is able to depend on other built
 }>
 <TabItem value="ts">
 
-```js
+```ts
 test.use({ baseURL: 'https://playwright.dev' })
 ```
 
@@ -503,7 +503,7 @@ Fixtures can also be overridden where the base fixture is completely replaced wi
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test as base } from '@playwright/test';
 
 export const test = base.extend({
