+++ nodejs/docs/test-parallel.mdx
@@ -50,7 +50,7 @@ In the configuration file:
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
 
@@ -104,7 +104,7 @@ Note that parallel tests are executed in separate worker processes and cannot sh
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test } from '@playwright/test';
 
 test.describe.parallel('suite', () => {
@@ -164,7 +164,7 @@ Setting in the configuration file:
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
