+++ nodejs/docs/test-pom.mdx
@@ -19,7 +19,7 @@ We will create a `PlaywrightDevPage` helper class to encapsulate common operatio
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright-dev-page.ts
 import { expect, Locator, Page } from '@playwright/test';
 
@@ -106,7 +106,7 @@ Now we can use the `PlaywrightDevPage` class in our tests.
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // example.spec.ts
 import { test, expect } from '@playwright/test';
 import { PlaywrightDevPage } from './playwright-dev-page';
