+++ nodejs/docs/test-reporters.mdx
@@ -30,7 +30,7 @@ For more control, you can specify reporters programmatically in the [configurati
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
 
@@ -72,7 +72,7 @@ You can use multiple reporters at the same time. For example  you can use`'list'
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
 
@@ -120,7 +120,7 @@ You can use different reporters locally and on CI. For example, using concise `'
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
 
@@ -164,7 +164,7 @@ You can use the built in `github` reporter to get automatic failure annotations 
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
 
@@ -216,7 +216,7 @@ npx playwright test --reporter=list
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
 
@@ -280,7 +280,7 @@ npx playwright test --reporter=line
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
 
@@ -341,7 +341,7 @@ npx playwright test --reporter=dot
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
 
@@ -383,18 +383,39 @@ JSON reporter produces an object with all information about the test run. It is 
 
 Most likely you want to write the JSON to a file. When running with `--reporter=json`, use `PLAYWRIGHT_JSON_OUTPUT_NAME` environment variable:
 
+<Tabs
+  groupId="bash-flavor"
+  defaultValue="bash"
+  values={[
+    {label: 'Bash', value: 'bash'},
+    {label: 'PowerShell', value: 'powershell'},
+    {label: 'Batch', value: 'batch'}
+  ]
+}>
+<TabItem value="bash">
+
 ```bash
-# Linux/macOS
 PLAYWRIGHT_JSON_OUTPUT_NAME=results.json npx playwright test --reporter=json,dot
+```
 
-# Windows with cmd.exe
-set PLAYWRIGHT_JSON_OUTPUT_NAME=results.json
-npx playwright test --reporter=json,dot
+</TabItem>
+<TabItem value="powershell">
 
-# Windows with PowerShell
+```powershell
 $env:PLAYWRIGHT_JSON_OUTPUT_NAME="results.json"
 npx playwright test --reporter=json,dot
 ```
+
+</TabItem>
+<TabItem value="batch">
+
+```batch
+set PLAYWRIGHT_JSON_OUTPUT_NAME=results.json
+npx playwright test --reporter=json,dot
+```
+
+</TabItem>
+</Tabs>
 
 In configuration file, pass options directly:
 
@@ -408,7 +429,7 @@ In configuration file, pass options directly:
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
 
@@ -442,18 +463,39 @@ JUnit reporter produces a JUnit-style xml report. It is usually used together wi
 
 Most likely you want to write the report to an xml file. When running with `--reporter=junit`, use `PLAYWRIGHT_JUNIT_OUTPUT_NAME` environment variable:
 
+<Tabs
+  groupId="bash-flavor"
+  defaultValue="bash"
+  values={[
+    {label: 'Bash', value: 'bash'},
+    {label: 'PowerShell', value: 'powershell'},
+    {label: 'Batch', value: 'batch'}
+  ]
+}>
+<TabItem value="bash">
+
 ```bash
-# Linux/macOS
 PLAYWRIGHT_JUNIT_OUTPUT_NAME=results.xml npx playwright test --reporter=junit,line
+```
 
-# Windows with cmd.exe
-set PLAYWRIGHT_JUNIT_OUTPUT_NAME=results.xml
-npx playwright test --reporter=junit,line
+</TabItem>
+<TabItem value="powershell">
 
-# Windows with PowerShell
+```powershell
 $env:PLAYWRIGHT_JUNIT_OUTPUT_NAME="results.xml"
 npx playwright test --reporter=junit,line
 ```
+
+</TabItem>
+<TabItem value="batch">
+
+```batch
+set PLAYWRIGHT_JUNIT_OUTPUT_NAME=results.xml
+npx playwright test --reporter=junit,line
+```
+
+</TabItem>
+</Tabs>
 
 In configuration file, pass options directly:
 
@@ -467,7 +509,7 @@ In configuration file, pass options directly:
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
 
@@ -509,7 +551,7 @@ You can create a custom reporter by implementing a class with some of the report
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { Reporter } from '@playwright/test/reporter';
 
@@ -577,7 +619,7 @@ Now use this reporter with [testConfig.reporter](./api/class-testconfig.mdx#test
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
