+++ nodejs/docs/test-retries.mdx
@@ -26,7 +26,7 @@ Consider the following snippet:
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test } from '@playwright/test';
 
 test.describe('suite', () => {
@@ -94,7 +94,7 @@ npx playwright test --retries=3
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // playwright.config.ts
 import { PlaywrightTestConfig } from '@playwright/test';
 
@@ -158,7 +158,7 @@ Consider the following snippet that uses `test.describe.serial`:
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { test } from '@playwright/test';
 
 test.describe.serial('suite', () => {
@@ -218,7 +218,7 @@ Playwright Test creates an isolated [Page] object for each test. However, if you
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // example.spec.ts
 
 import { test, Page } from '@playwright/test';
