+++ nodejs/docs/test-snapshots.mdx
@@ -17,7 +17,7 @@ Playwright Test includes the ability to produce and visually compare screenshots
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // example.spec.ts
 import { test, expect } from '@playwright/test';
 
@@ -89,7 +89,7 @@ Playwright Test uses the [pixelmatch](https://github.com/mapbox/pixelmatch) libr
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // example.spec.ts
 import { test, expect } from '@playwright/test';
 
@@ -127,7 +127,7 @@ If you'd like to share the default value among all the tests in the project, you
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { PlaywrightTestConfig } from '@playwright/test';
 const config: PlaywrightTestConfig = {
   expect: {
@@ -165,7 +165,7 @@ Here we compare text content against the reference.
 }>
 <TabItem value="ts">
 
-```js
+```ts
 // example.spec.ts
 import { test, expect } from '@playwright/test';
