+++ nodejs/docs/trace-viewer.mdx
@@ -31,7 +31,7 @@ Set the `trace: 'on-first-retry'` option in the test configuration file. This wi
 }>
 <TabItem value="ts">
 
-```js
+```ts
 import { PlaywrightTestConfig } from '@playwright/test';
 const config: PlaywrightTestConfig = {
   retries: 1,
@@ -62,7 +62,7 @@ module.exports = config;
 </TabItem>
 <TabItem value="library">
 
-```js
+```library
 const browser = await chromium.launch();
 const context = await browser.newContext();
