+++ python/docs/inspector.mdx
@@ -20,18 +20,39 @@ Playwright Inspector is a GUI tool that helps authoring and debugging Playwright
 There are several ways of opening Playwright Inspector:
 - Set the `PWDEBUG` environment variable to run your scripts in debug mode. This configures Playwright for debugging and opens the inspector.
 
+  <Tabs
+    groupId="bash-flavor"
+    defaultValue="bash"
+    values={[
+      {label: 'Bash', value: 'bash'},
+      {label: 'PowerShell', value: 'powershell'},
+      {label: 'Batch', value: 'batch'}
+    ]
+  }>
+  <TabItem value="bash">
+  
   ```bash
-  # Linux/macOS
   PWDEBUG=1 pytest -s
+  ```
   
-  # Windows with cmd.exe
-  set PWDEBUG=1
-  pytest -s
+  </TabItem>
+  <TabItem value="powershell">
   
-  # Windows with PowerShell
+  ```powershell
   $env:PWDEBUG=1
   pytest -s
   ```
+  
+  </TabItem>
+  <TabItem value="batch">
+  
+  ```batch
+  set PWDEBUG=1
+  pytest -s
+  ```
+  
+  </TabItem>
+  </Tabs>
 
   Additional useful defaults are configured when `PWDEBUG=1` is set:
   - Browsers launch in the headed mode
