+++ src/format_java.js
@@ -17,6 +17,7 @@
 //@ts-check
 
 const Documentation = require('./documentation');
+const { generateTabGroup } = require('./format_utils');
 const { toTitleCase, renderJSSignature } = require('./generator');
 /** @typedef {import('./generator').GeneratorFormatter} GeneratorFormatter */
 
@@ -24,6 +25,9 @@ const { toTitleCase, renderJSSignature } = require('./generator');
  * @implements {GeneratorFormatter}
  */
 class JavaFormatter {
+  constructor() {
+    this.lang = 'java';
+  }
   formatMember(member) {
     let text;
     let args = [];
@@ -127,7 +131,7 @@ class JavaFormatter {
       if (n.text === 'extends: [Error]')
         n.text = 'extends: [PlaywrightException]';
     });
-    return spec;
+    return generateTabGroup(spec, this.lang, 'bash');
   }
 }
